issue: record optimization ideas for the future
The version I have now is easy to reason about. First goal is to get the CPU
working, and the I can come back and optimize it.  However, that is a long
ways off and I won't remember the issue unit by that point.  So, record the
ideas I have now and then I can potentially use them to make it faster later.