pipeline: move backpressure from fifo to renamer
To make good use of reservation stations, we need to refill them
quickly once they are committed. Trying to use tokens in the FIFO
for this adds too many cycles of latency, wasting FIFO slots.

This commit puts the back-pressure in the renamer, which means
that a reservation station is only wasted for 2 cycles:
  useful data in station     (r_stat_reg*)
  commit decision            (r_done_regx => commit_stb)
  station readiness check    (r_ren_stb => s_ren_done_a)