regfile: a better mux system
This approach tries to keep the critical fast execute path short.
It is also more portable as it makes no assumptions about synthesis.

The cost is that r_mux_idx_[ab] have mega-fan-in, though a fairly shallow
circuit.  However, as there are so few index bits, it seems this cost is
negligible. The fitter can move it to the center of a big web.