opa: best IPC seems to be when there are stations = 10-12 * max-throughput
For a 4-way 2+1 processor, the max throughput is 3, so ideally 36 stations.
However, 28 stations is the maximum we can fit before the register file goes over 64.

For a 4-way 2+2 processor, the max throughput is 4, so go with 44 stations.

Even for a "small" 2-way 1+1 processor, we want 22 stations, but again there is a nice
break point where 18 stations means you can schedule 1hot decode stuff in two layers.

For the truly budget CPU, use 1-way with 6 stations to support decode in one LUT.