opa: use 1-hot instruction decoding style
Previously, we packed all decoding into one 8-bit register.

Now that same structure takes 24 bits. However, the advantage is that
the synthesis tool can now tell the bits appart all the way from the
decode stage to the execute stage. So, if there are no instructions in
the ISA that use all 4 cases of the LUT operation, for example, those
can be merged all the way later in the execute stage.

The cost for a 4-decode 3-issue LM32 is:
  +100 LUTs in decode (including 50 new regs)
   +50 LUTs in regfile
=> a 2% increase in area.