issue: fixed two load cam bugs; shifting wrong axis and leaking an issue
r_alias_cam was shifted on the wrong axis. easily fixed.
then the cpu was getting wedged on heavy load alias workloads.
the added assertion detects a leaked instruction.

the bug was failure to consider r_{fast,slow}_issue as needing to be
cleaned up just like r_issue. s_issue is the union of the three, so
they all need to be consistent.