l1d: define misaligned access to be intraword rotation
This solves the possibility that a load aliases a store that has
a different word address. I will still need to fault for RV32M, but
for LM32 at least this 'undefined' behaviour does not cause behaviour
that depends on scheduler order / luck.

The new implementation also seems to be 100 ALMs smaller per way.
Not sure why, but I'll take it!